named-functions
If you use a named function writes this first when the event appears on the dom. There are no parentheses on the function name. This means that arguments can't passed t. The text input whose id attribute has a value of user, is the function. The dom element node is stored in a variable. The text input is placed into a variable called elUsername.