.html-.text()-methods
These methods both retrieve and update the content of elements.
In these examples the selector returns the <ul> element. The .html() method gets all the HTML inside the <li> elements. This is then added to the end of the selection, i.e. after the existing <li>elements.
The .text() method gets the text from all of the <ul> element's children. This is then attached to the end of the selection, i.e. after the existing <ul> element.