setting-and-changing-css-rules
The .css() can be used to select and update the css properties of elements. The script checks what the background colour of the first list item is when the page loads and then writes it after the list.
Next, it updates several css properties in all list items using the same .css() method with object literal notation.