event-handlers-for-delegating-events
The event handlers run when the user clicks or mouseover items on the list, except for the last item.
It writes out the content of the element the user interacted with. A status message (using the data property)
and the event type. The info passed in the data property here uses object literal notation (so it can handle
multiple properties.